Backend/requirements.txt
fastapi
uvicorn
xarray
netCDF4
numpy
pandas
faiss-cpu
pydantic
python-dotenv
transformers
torch
huggingface-hub
sentence-transformers
sentencepiece